# lib/api-zod/src/generated/types/createPage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Musholla Nurul Iman API
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePage {
  title: string;
  slug: string;
  content: string;
  isPublished: boolean;
}