# lib/api-zod/src/generated/types/siteSettings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Musholla Nurul Iman API
 * OpenAPI spec version: 0.1.0
 */

export interface SiteSettings {
  siteName: string;
  tagline: string;
  description: string;
  address: string;
  phone: string;
  email: string;
  facebook?: string | null;
  twitter?: string | null;
  instagram?: string | null;
  youtube?: string | null;
  logoUrl?: string | null;
  heroImageUrl?: string | null;
  heroTitle: string;
  heroSubtitle: string;
}